+++ OT/src/ot/ukirt/tpe/TpeDichroicFeature.java
@@ -97,8 +97,8 @@ public class TpeDichroicFeature extends TpeImageFeature
 		double[] xpoints = _fovAreaPD.xpoints ;
 		double[] ypoints = _fovAreaPD.ypoints ;
 
-		double x = ( double )fii.baseScreenPos.x ;
-		double y = ( double )fii.baseScreenPos.y ;
+		double x = fii.baseScreenPos.x;
+		double y = fii.baseScreenPos.y;
 
 		double wl = fii.pixelsPerArcsec * FOV_WIDTHL ;
 		double wr = fii.pixelsPerArcsec * FOV_WIDTHR ;
